Set a minimum pin on psycopg2
There is a bug in psycopg2==2.4.5 where it raises a `TypeError` if the
`name` parameter passed to the `connection.cursor` method is None,
despite the fact the default argument is `None`.  When queries creates
a new Session it passes `None` to the cursor function and explodes.
This behavior is fixed in versions later than 2.5.1.